@@ -4,9 +4,12 @@
   "private": true,
   "dependencies": {
     "@date-io/date-fns": "1.x",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
     "@material-ui/core": "^4.7.2",
     "@material-ui/icons": "^4.5.1",
     "@material-ui/pickers": "^3.2.10",
+    "@mui/material": "^5.2.8",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
@@ -14,24 +17,24 @@
     "date-fns": "^2.14.0",
     "eslint-config-prettier": "^6.10.1",
     "material-ui": "^0.20.2",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0",
     "react-draggable": "^4.2.0",
     "react-dropzone": "^10.2.1",
     "react-infinite-scroll-component": "^5.0.4",
     "react-player": "^1.14.2",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.0",
+    "react-scripts": "^4.0.0",
     "react-simple-keyboard": "^2.1.82",
     "react-use": "^13.12.2",
     "react-vis": "^1.11.7",
     "recharts": "^2.0.0-beta.1",
     "socket.io-client": "^2.3.0",
     "store": "^2.0.12",
-    "styled-components": "^4.4.1"
+    "styled-components": "^5.3.3"
   },
   "devDependencies": {
-    "eslint": "^6.8.0",
+    "eslint": "^7.11.0",
     "eslint-loader": "^3.0.3",
     "prettier": "^1.19.1"
   },
