@@ -3,22 +3,27 @@ import React, { useState } from 'react';
 import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom';
 
 // ui
-import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
+import { createTheme, MuiThemeProvider } from '@material-ui/core/styles';
 
 // hook
 import { CommonContext } from './context/CommonContext';
 import { useLocalStorageSetState } from './common/CommonHooks';
 
 // page
+import Main from './pages/Main/';
 import Auth from './pages/Auth/';
 import Terms from './pages/Terms/';
 import MyVote from './pages/MyVote/';
-import AboutMe from './pages/AboutMe/';
+import AboutTeam from './pages/AboutTeam/';
 import NotFound from './pages/NotFound/';
 import MainVote from './pages/MainVote/';
 import ContactUs from './pages/ContactUs/';
 import CreateVote from './pages/CreateVote/';
 import SearchVote from './pages/SearchVote/';
+import Ask from './pages/Ask/';
+import Community from './pages/Community/';
+import Admin from './pages/Admin/';
+import MyFarm from './pages/MyFarm/';
 
 // css
 // import './index.css';
@@ -31,7 +36,7 @@ const serverUrlBase = `http://${HOST}`;
 const serverImgUrl = `https://ssafy-viba-s3.s3.ap-northeast-2.amazonaws.com/public/`;
 
 /// theme
-const theme = createMuiTheme({
+const theme = createTheme({
   typography: {
     fontFamily: ['Noto Sans KR'].join(','),
     button: {
@@ -71,6 +76,7 @@ const App = () => {
   const [drawerOpen, setDrawerOpen] = useState(false);
   const [userDialogIndex, setUserDialogIndex] = useState(0);
   const [isShowKeyborad, setIsShowKeyborad] = useState(false);
+  const [isSignUp, setIsSignUp] = useState('SignIn');
   const [signDialogOpen, setSignDialogOpen] = useState(false);
   const [infoDialogOpen, setInfoDetailDialogOpen] = useState(false);
   const [userDialogOpen, setUserDetailDialogOpen] = useState(false);
@@ -98,21 +104,27 @@ const App = () => {
         isShowKeyborad,
         setIsShowKeyborad,
         defaultThumbnailImage,
+        isSignUp,
+        setIsSignUp,
       }}
     >
       <MuiThemeProvider theme={theme}>
         <BrowserRouter>
           <Switch>
-            <Route exact path="/" component={MainVote} />
+            <Route exact path="/" component={Main} />
             <Route exact path="/MainVote" component={MainVote} />
             <Route exact path="/Auth" component={Auth} />
             <Route exact path="/Terms" component={Terms} />
             <Route exact path="/MyVote" component={MyVote} />
-            <Route exact path="/AboutMe" component={AboutMe} />
+            <Route exact path="/AboutTeam" component={AboutTeam} />
             <Route exact path="/ContactUs" component={ContactUs} />
             <Route exact path="/SearchVote" component={SearchVote} />
             <Route exact path="/not-found" component={NotFound} />
             <Route exact path="/CreateVote" component={CreateVote} />
+            <Route exact path="/Ask" component={Ask} />
+            <Route exact path="/Community" component={Community} />
+            <Route exact path="/MyFarm" component={MyFarm} />
+            <Route exact path="/Admin" component={Admin} />
             <Redirect to="/not-found" />
           </Switch>
         </BrowserRouter>
