@@ -16,16 +16,30 @@ import {
   Divider,
   TextField,
 } from '@material-ui/core';
+
+// Input 안에 icon 넣을 거라면
+import InputAdornment from '@material-ui/core/InputAdornment';
+
 import Wrapper from './styles';
 
 import userData from './dump.json';
 
+// 이메일 체크 정규식
 const regExp = /^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*\.[a-zA-Z]{2,3}$/i;
+
+// 비밀번호 체크 정규식
+// var regex = /^.*(?=^.{8,15}$)(?=.*\d)(?=.*[a-zA-Z])(?=.*[!@#$%^&+=]).*$/;
+// 특수 / 문자 / 숫자 포함 형태 (8~15)
+
+// 핸드폰 번호 체크 정규식
+// var regExp = /^\d{3}-\d{3,4}-\d{4}$/;
 
 const DialogTitleComponent = () => {
   return (
     <Wrapper>
-      <h1 className="dialog-title-component">{'Logo'}</h1>
+      <h1 className="dialog-title-component">
+        <img className="logo_img" src="images/ssug_green.png" alt="logo" />
+      </h1>
     </Wrapper>
   );
 };
@@ -64,10 +78,10 @@ const SignInSection01 = () => {
       return;
     }
 
-    if (!regExp.test(id)) {
-      alert('The email format is invalid.');
-      return;
-    }
+    // if (!regExp.test(id)) {
+    //   alert('The email format is invalid.');
+    //   return;
+    // }
 
     let respone = [];
     let hashPassword = '';
@@ -81,6 +95,8 @@ const SignInSection01 = () => {
     }
 
     setUser({ ...userData });
+    // 지금은 dump지만 / 나중엔 signInUserData?
+
     setSignDialogOpen(false);
     setIsSignUp('SignIn');
 
@@ -89,15 +105,22 @@ const SignInSection01 = () => {
 
   useEffect(() => {
     console.log({ user });
-
-    if (signInUserData.id !== '' && signInUserData.email !== '') {
+    if (signInUserData.id !== '' && signInUserData.password !== '') {
       setDisabled(false);
     }
 
-    if (signInUserData.id === '' || signInUserData.email === '') {
+    if (signInUserData.id === '' || signInUserData.password === '') {
       setDisabled(true);
     }
-  }, [signInUserData.id, signInUserData.email, user]);
+  }, [signInUserData.id, signInUserData.password, user]);
+  //   if (signInUserData.id !== '' && signInUserData.email !== '') {
+  //     setDisabled(false);
+  //   }
+
+  //   if (signInUserData.id === '' || signInUserData.email === '') {
+  //     setDisabled(true);
+  //   }
+  // }, [signInUserData.id, signInUserData.email, user]);
 
   return (
     <Wrapper>
@@ -113,7 +136,7 @@ const SignInSection01 = () => {
           <TextField
             required
             id="outlined-required"
-            label="Email"
+            label="아이디"
             className="text-field"
             defaultValue={signInUserData.id}
             variant="outlined"
@@ -121,6 +144,14 @@ const SignInSection01 = () => {
             onChange={OnChangeHandler('id')}
             onFocus={event => {
               setIsShowKeyborad(true);
+              // 아이콘 양식
+              // InputProps={{
+              //   startAdornment: (
+              //     <InputAdornment position="start">
+              //       <PersonIcon />
+              //     </InputAdornment>
+              //   ),
+              // }}
             }}
           />
         </Grid>
@@ -128,7 +159,7 @@ const SignInSection01 = () => {
           <TextField
             required
             id="outlined-password-input"
-            label="Password"
+            label="비밀번호"
             className="text-field"
             type="password"
             autoComplete="current-password"
@@ -147,14 +178,14 @@ const SignInSection01 = () => {
             disabled={disabled}
             // disabled={false}
             fullWidth={true}
-            color="primary"
+            // color="primary"
             onClick={onSignInHandler}
             className="grid-item-button"
           >
-            login
+            로그인
           </Button>
         </Grid>
-        <Grid item xs={12} className="grid-item">
+        {/* <Grid item xs={12} className="grid-item">
           <Grid
             container
             direction="row"
@@ -176,7 +207,7 @@ const SignInSection01 = () => {
               <Divider />
             </Grid>
           </Grid>
-        </Grid>
+        </Grid> */}
 
         <Grid item xs={12}>
           <Grid container direction="row" justify="center" alignItems="center">
@@ -185,7 +216,7 @@ const SignInSection01 = () => {
               onClick={onClickHandler}
             >
               <Typography className="grid-item-typography3">
-                {'Forgot Password?'}
+                {'비밀번호 찾기'}
               </Typography>
             </IconButton>
           </Grid>
@@ -209,7 +240,7 @@ const SignInSection02 = () => {
         justify="center"
         alignItems="center"
         spacing={1}
-        className="grid2"
+        className="grid"
       >
         <Grid item xs={12}>
           <Button
@@ -217,7 +248,7 @@ const SignInSection02 = () => {
             onClick={onClickHandler}
             className="grid2-item-button"
           >
-            {`Don't have an account?`}
+            {`회원가입`}
           </Button>
         </Grid>
       </Grid>
@@ -247,11 +278,13 @@ const SignInGroupComponent = () => {
   );
 };
 
+////////////////////////////////////////////////////////////////////////////////
+
 const SignUpSection01 = () => {
   return (
     <Wrapper>
       <Typography align="center" className="sign-up1">
-        Sign up to see friends' photos and videos.
+        쑥쑥에 가입해서 ... 해보세요
       </Typography>
     </Wrapper>
   );
@@ -266,16 +299,22 @@ const SignUpSection02 = () => {
 
   const OnChangeHandler = name => e => {
     if (
-      signUpUserData.name !== '' &&
       signUpUserData.id !== '' &&
+      signUpUserData.password !== '' &&
+      signUpUserData.passwordConfirmation !== '' &&
+      signUpUserData.name !== '' &&
+      signUpUserData.nickname !== '' &&
       signUpUserData.email !== ''
     ) {
       setDisabled(false);
     }
 
     if (
-      signUpUserData.name === '' ||
       signUpUserData.id === '' ||
+      signUpUserData.password === '' ||
+      signUpUserData.passwordConfirmation === '' ||
+      signUpUserData.name === '' ||
+      signUpUserData.nickname === '' ||
       signUpUserData.email === ''
     ) {
       setDisabled(true);
@@ -285,14 +324,28 @@ const SignUpSection02 = () => {
   };
 
   const onSignUpHandler = async () => {
-    var { name, id, password } = signUpUserData;
+    var {
+      id,
+      password,
+      passwordConfirmation,
+      name,
+      nickname,
+      email,
+    } = signUpUserData;
 
-    if (name === '' || id === '' || password === '') {
-      alert('You need both email and password and username.');
+    if (
+      id === '' ||
+      password === '' ||
+      passwordConfirmation === '' ||
+      name === '' ||
+      nickname === '' ||
+      email === ''
+    ) {
+      alert('You need 문구는 수정해야 ! both email and password and username.');
       return;
     }
 
-    if (!regExp.test(id)) {
+    if (!regExp.test(email)) {
       alert('The email format is invalid.');
       return;
     }
@@ -308,6 +361,7 @@ const SignUpSection02 = () => {
       console.log('PPAP: signInHandler -> error', error);
     }
 
+    // 여기 잠시만...??
     var body = {
       id: id,
       name: name,
@@ -315,13 +369,14 @@ const SignUpSection02 = () => {
     };
     console.log('PPAP: signUpHandler -> body', body);
 
-    //
-
     setIsSignUp('SignIn');
     setSignUpUserData({
       id: '',
-      name: '',
       password: '',
+      passwordConfirmation: '',
+      name: '',
+      nickname: '',
+      email: '',
     });
   };
 
@@ -333,13 +388,14 @@ const SignUpSection02 = () => {
         justify="center"
         alignItems="center"
         spacing={1}
-        style={{ marginLeft: 4 }}
+        // style={{ marginLeft: 4 }}
+        className="grid"
       >
         <Grid item xs={12} className="sign-up-grid">
           <TextField
             required
             id="outlined-required"
-            label="Email"
+            label="아이디"
             defaultValue={signUpUserData.id}
             className="text-field"
             variant="outlined"
@@ -348,11 +404,41 @@ const SignUpSection02 = () => {
             onChange={OnChangeHandler('id')}
           />
         </Grid>
+        <Grid item xs={12} className="sign-up-grid-item2">
+          <TextField
+            required
+            id="outlined-password-input"
+            label="비밀번호"
+            className="text-Field"
+            type="password"
+            autoComplete="current-password"
+            defaultValue={signUpUserData.password}
+            variant="outlined"
+            placeholder=""
+            fullWidth={true}
+            onChange={OnChangeHandler('password')}
+          />
+        </Grid>
+        <Grid item xs={12} className="sign-up-grid-item2">
+          <TextField
+            required
+            id="outlined-password-input"
+            label="비밀번호확인"
+            className="text-Field"
+            type="password"
+            autoComplete="current-password"
+            defaultValue={signUpUserData.passwordConfirmation}
+            variant="outlined"
+            placeholder=""
+            fullWidth={true}
+            onChange={OnChangeHandler('password')}
+          />
+        </Grid>
         <Grid item xs={12} className="sign-up-grid-item1">
           <TextField
             required
             id="outlined-required"
-            label="User Name"
+            label="이름"
             defaultValue={signUpUserData.name}
             className="text-field"
             variant="outlined"
@@ -361,19 +447,30 @@ const SignUpSection02 = () => {
             onChange={OnChangeHandler('name')}
           />
         </Grid>
-        <Grid item xs={12} className="sign-up-grid-item2">
+        <Grid item xs={12} className="sign-up-grid-item1">
           <TextField
             required
-            id="outlined-password-input"
-            label="Password"
-            className="textField"
-            type="password"
-            autoComplete="current-password"
-            defaultValue={signUpUserData.password}
+            id="outlined-required"
+            label="별명"
+            defaultValue={signUpUserData.nickname}
+            className="text-field"
             variant="outlined"
-            placeholder="비밀번호"
+            placeholder=""
             fullWidth={true}
-            onChange={OnChangeHandler('password')}
+            onChange={OnChangeHandler('name')}
+          />
+        </Grid>
+        <Grid item xs={12} className="sign-up-grid">
+          <TextField
+            required
+            id="outlined-required"
+            label="이메일"
+            defaultValue={signUpUserData.email}
+            className="text-field"
+            variant="outlined"
+            placeholder=""
+            fullWidth={true}
+            onChange={OnChangeHandler('id')}
           />
         </Grid>
         <Grid item xs={12} className="sign-up-grid-item3">
@@ -381,21 +478,23 @@ const SignUpSection02 = () => {
             variant="contained"
             disabled={disabled}
             fullWidth={true}
-            color="primary"
+            // color="primary"
             onClick={onSignUpHandler}
+            className="grid-item-button"
             style={{
               fontSize: 14,
               fontFamily: 'Noto Sans KR',
               fontWeight: 500,
             }}
           >
-            Sign up
+            회원가입
           </Button>
         </Grid>
         <Grid item xs={12} className="sign-up-grid-item4">
           <Typography align="center" className="sign-up-grid-item4-typography">
-            By signing up, you agree to ssafy's terms, <b>data policy</b> and{' '}
-            <b>cookie policy.</b>
+            쑥쑥 시스템에 가입함으로써
+            <br /> 귀하는 저희의 약관과 <b>데이터 및 쿠키 정책</b>에 동의하시게
+            됩니다.
           </Typography>
         </Grid>
       </Grid>
@@ -415,20 +514,20 @@ const SignUpSection03 = () => {
           spacing={1}
           className="sign-up3-grid-item"
         >
-          <Grid item xs={5}>
+          {/* <Grid item xs={5}>
             <Divider />
-          </Grid>
+          </Grid> */}
 
-          <Grid item xs={2}>
+          {/* <Grid item xs={2}>
             <Typography
               align="center"
               className="sign-up3-grid-item-typography"
             >
               or
             </Typography>
-          </Grid>
+          </Grid> */}
 
-          <Grid item xs={5}>
+          <Grid item xs={12}>
             <Divider />
           </Grid>
         </Grid>
@@ -453,19 +552,19 @@ const SignUpSection04 = () => {
         alignItems="center"
         className="sign-up4-grid"
       >
-        <Grid item xs={2} />
-        <Grid item xs={5}>
+        <Grid item xs={1} />
+        <Grid item xs={6}>
           <Typography align="center" className="sign-up4-grid-item-typography">
-            {'Do you have an account?'}
+            {'이미 계정이 있습니까?'}
           </Typography>
         </Grid>
-        <Grid item xs={1}>
+        <Grid item xs={3}>
           <Button
             fullWidth={true}
             onClick={onClickHandler}
             className="sign-up4-grid-item-button"
           >
-            {'login'}
+            {'로그인하기'}
           </Button>
         </Grid>
       </Grid>
@@ -501,6 +600,8 @@ const SignUpGroupComponent = () => {
   );
 };
 
+////////////////////////////////////////////////////////////////////////////////
+
 // ForgotPw
 const ForgotPwGroupComponent = () => {
   const { serverUrl } = useContext(CommonContext);
@@ -513,16 +614,15 @@ const ForgotPwGroupComponent = () => {
     setIsSignUp(whichGroup);
   };
   const sendSearchWordHandler = async searchWord => {
-    if (!regExp.test(searchWord)) {
-      alert('The email format is invalid.');
-      return;
-    } else {
-      alert('Not implemented yet.');
-
-      // setRecoverPwUserData({ ...recoverPwUserData, email: searchWord });
-      // alert('Authentication code has been sent to you by email');
-      // setIsSignUp('RecoverPw');
-    }
+    // if (!regExp.test(searchWord)) {
+    //   alert('The email format is invalid.');
+    //   return;
+    // } else {
+    //   alert('Not implemented yet.');
+    //   // setRecoverPwUserData({ ...recoverPwUserData, email: searchWord });
+    //   // alert('Authentication code has been sent to you by email');
+    //   // setIsSignUp('RecoverPw');
+    // }
   };
   return (
     <Wrapper>
@@ -534,12 +634,12 @@ const ForgotPwGroupComponent = () => {
         spacing={1}
         className="forgot-pw"
       >
-        <h2>Having trouble signing in?</h2>
+        <h2>비밀번호를 잊어버리셨나요?</h2>
         <h3>
           Enter the user ID and the verification code will be sent to the
           registered email.
         </h3>
-        <input type="text" placeholder="User ID" ref={inputRef} />
+        <input type="text" placeholder="아이디" ref={inputRef} />
         <button
           type="button"
           className="btn-link"
@@ -550,7 +650,7 @@ const ForgotPwGroupComponent = () => {
           Send Login Link
         </button>
         <h4 className="divider">
-          <span>or</span>
+          <span>계정이 없다면 바로 가입하세요!</span>
         </h4>
         <h5
           className="btn-to-sign-up"
@@ -558,7 +658,7 @@ const ForgotPwGroupComponent = () => {
             onClickHandler(`SignUp`);
           }}
         >
-          Create new account
+          회원가입
         </h5>
         <button
           type="button"
@@ -567,8 +667,11 @@ const ForgotPwGroupComponent = () => {
             onClickHandler(`SignIn`);
           }}
         >
-          Back to login
+          로그인하기
         </button>
+        <Grid item xs={12}>
+          <div>&nbsp;</div>
+        </Grid>
       </Grid>
     </Wrapper>
   );
@@ -684,11 +787,13 @@ const RecoverPwGroupComponent = () => {
   );
 };
 
+////////////////////////////////////////////////////////////////////////////////
+
 const ResponsiveDialogSign = () => {
   const fullScreen = useMediaQuery(theme => theme.breakpoints.down('xs'));
   let history = useHistory();
 
-  const { signDialogOpen, setSignDialogOpen, serverImgUrl } = useContext(
+  const { signDialogOpen, setSignDialogOpen, serverImgUrl, isSignUp, setIsSignUp } = useContext(
     CommonContext,
   );
 
@@ -698,7 +803,7 @@ const ResponsiveDialogSign = () => {
     history.goBack();
   };
 
-  const [isSignUp, setIsSignUp] = useState('SignIn');
+  // const [isSignUp, setIsSignUp] = useState('SignIn');
   const [signInUserData, setSignInUserData] = useState({
     id: '',
     password: '',
@@ -735,18 +840,18 @@ const ResponsiveDialogSign = () => {
         aria-labelledby="responsive-dialog-title"
         PaperProps={{
           style: {
-            backgroundColor: 'white',
+            backgroundColor: '#ffffff',
             boxShadow: 'none',
           },
         }}
-        BackdropProps={{
-          style: {
-            boxShadow: 'none',
-            backgroundImage: `url(${serverImgUrl}thumb-1920-731946.jpg)`,
-            backgroundSize: 'cover',
-            filter: 'brightness(0.4)',
-          },
-        }}
+        // BackdropProps={{
+        //   style: {
+        //     boxShadow: 'none',
+        //     backgroundImage: `url(${serverImgUrl}thumb-1920-731946.jpg)`,
+        //     backgroundSize: 'cover',
+        //     filter: 'brightness(0.4)',
+        //   },
+        // }}
       >
         <Grid container direction="row" justify="center" alignItems="center">
           <Grid item xs={12}>
