@@ -31,6 +31,16 @@ const Wrapper = styled.div`
   & .up-cancel-fab {
     border-radius: 30px;
     margin-right: 12px;
+
+    border-color: #a4a4a4;
+    color:#a4a4a4;
+    background: #ffffff;
+    box-shadow: none;
+
+    &:hover{
+      border-color: #000000;
+      color:#000000;
+    }
   }
   & .bg-unset {
     &:hover {
@@ -69,7 +79,7 @@ const Wrapper = styled.div`
       }
     }
   }
-  @media (max-width: 960px) {
+  @media (max-width: 1100px) {
     & .drawer {
       display: block;
     }
